move App out of the '/routes' folder
